requirements.txt
phidata 
phi
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
Gemini
google-generativeai
python-multipart